+++ src/QuartetConsensus.cpp
@@ -55,7 +55,7 @@ inline int32 quartet_index(int16 a, int16 b, int16 c, int16 d,
        + chosen4;
 }
 
-inline int32 n_quartets(int16 n_tips) {
+static inline int32 n_quartets(int16 n_tips) {
   return hyp_num[n_tips - 3];
 }